+++ lib/constants/translations/app_strings.dart
@@ -74,10 +74,17 @@ class AppStrings {
   String get startGame => getValue('startGame');
   String get fastTrackTo => getValue('fastTrackTo');
   String get loadGame => getValue('loadGame');
+  String get continueGame => getValue('continueGame');
+  String get overwriteConfirmation => getValue('overwriteConfirmation');
+  String get continueText => getValue('continueText');
+  String get round => getValue('round');
+  String get gold => getValue('gold');
+  String get flameMana => getValue('flameMana');
 
   // Between rounds
   String get startRound => getValue('startRound');
   String get saveGame => getValue('saveGame');
+  String get saveAndQuit => getValue('saveAndQuit');
   String get enterShop => getValue('enterShop');
 
   // Shop