+++ lib/core/flame/components/hud/game_selection_internal/game_selection_hud_state.dart
@@ -1,0 +1,1 @@
+enum GameSelectionHudState { menu, saveOverwriteConfirmation }