+++ lib/core/flame/components/masonry/player_base.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:defend_your_flame/constants/bounding_constants.dart';
 import 'package:defend_your_flame/constants/debug_constants.dart';
+import 'package:defend_your_flame/constants/theming_constants.dart';
 import 'package:defend_your_flame/core/flame/components/entities/entity.dart';
 import 'package:defend_your_flame/core/flame/components/entities/npcs/blacksmith.dart';
 import 'package:defend_your_flame/core/flame/components/masonry/fire_pit.dart';
@@ -10,6 +11,7 @@ import 'package:defend_your_flame/core/flame/components/masonry/totems/attack_to
 import 'package:defend_your_flame/core/flame/components/masonry/walls/wall.dart';
 import 'package:defend_your_flame/core/flame/worlds/main_world.dart';
 import 'package:defend_your_flame/core/flame/worlds/main_world_state.dart';
+import 'package:defend_your_flame/core/storage/saves/game_save.dart';
 import 'package:flame/components.dart';
 import 'package:flutter/material.dart';
 
@@ -108,7 +110,7 @@ class PlayerBase extends PositionComponent with HasWorldReference<MainWorld>, Ha
     if (DebugConstants.drawBaseArea) {
       // This is used for debugging
       var relativeRect = _innerBaseRect.translate(-position.x, -position.y);
-      canvas.drawRect(relativeRect, DebugConstants.debugPaint);
+      canvas.drawRect(relativeRect, ThemingConstants.debugPaint);
     }
   }
 
@@ -133,5 +135,11 @@ class PlayerBase extends PositionComponent with HasWorldReference<MainWorld>, Ha
   void addFlameMana(int flameMana, Vector2 position) {
     _flameMana += flameMana;
     world.effectManager.addFlameText(flameMana, position);
+  }
+
+  void loadFromSave(GameSave save) {
+    _gold = save.currentGold;
+    _flameMana = save.currentFlameMana;
+    _wall.overrideHealth(save.currentWallHealth);
   }
 }