+++ lib/core/flame/managers/round_manager.dart
@@ -11,13 +11,6 @@ class RoundManager extends Component with HasWorldReference<MainWorld> {
     world.worldStateManager.changeState(MainWorldState.playing);
   }
 
-  void resetGame() {
-    world.playerBase.reset();
-    world.entityManager.clearEntities();
-    world.shopManager.resetPurchases();
-    _currentRound = 0;
-  }
-
   void overrideRound(int round) {
     _currentRound = round;
   }